Synchronize Repo

+++ wo/cli/plugins/clean.py
@@ -1,13 +1,14 @@
 """Clean Plugin for WordOps."""
 
-from wo.core.shellexec import WOShellExec
-from wo.core.aptget import WOAptGet
-from wo.core.services import WOService
-from wo.core.logging import Log
-from cement.core.controller import CementBaseController, expose
-from cement.core import handler, hook
 import os
-import urllib.request
+import requests
+
+from cement.core.controller import CementBaseController, expose
+
+from wo.core.aptget import WOAptGet
+from wo.core.logging import Log
+from wo.core.services import WOService
+from wo.core.shellexec import WOShellExec
 
 
 def wo_clean_hook(app):
@@ -20,14 +21,12 @@ class WOCleanController(CementBaseController):
         stacked_on = 'base'
         stacked_type = 'nested'
         description = (
-            'Clean NGINX FastCGI cache, Opcache, Memcached, Redis Cache')
+            'Clean NGINX FastCGI cache, Opcache, Redis Cache')
         arguments = [
             (['--all'],
                 dict(help='Clean all cache', action='store_true')),
             (['--fastcgi'],
                 dict(help='Clean FastCGI cache', action='store_true')),
-            (['--memcached'],
-                dict(help='Clean MemCached', action='store_true')),
             (['--opcache'],
                 dict(help='Clean OpCache', action='store_true')),
             (['--redis'],
@@ -37,22 +36,19 @@ class WOCleanController(CementBaseController):
 
     @expose(hide=True)
     def default(self):
-        if (not (self.app.pargs.all or self.app.pargs.fastcgi or
-                 self.app.pargs.memcached or self.app.pargs.opcache or
-                 self.app.pargs.redis)):
+        pargs = self.app.pargs
+        if ((not pargs.all) and (not pargs.fastcgi) and
+                (not pargs.opcache) and (not pargs.redis)):
             self.clean_fastcgi()
-        if self.app.pargs.all:
-            self.clean_memcached()
+        if pargs.all:
             self.clean_fastcgi()
             self.clean_opcache()
             self.clean_redis()
-        if self.app.pargs.fastcgi:
+        if pargs.fastcgi:
             self.clean_fastcgi()
-        if self.app.pargs.memcached:
-            self.clean_memcached()
-        if self.app.pargs.opcache:
+        if pargs.opcache:
             self.clean_opcache()
-        if self.app.pargs.redis:
+        if pargs.redis:
             self.clean_redis()
 
     @expose(hide=True)
@@ -65,44 +61,45 @@ class WOCleanController(CementBaseController):
             Log.info(self, "Redis is not installed")
 
     @expose(hide=True)
-    def clean_memcached(self):
-        try:
-            if(WOAptGet.is_installed(self, "memcached")):
-                WOService.restart_service(self, "memcached")
-                Log.info(self, "Cleaning MemCached")
-            else:
-                Log.info(self, "Memcached not installed")
-        except Exception as e:
-            Log.debug(self, "{0}".format(e))
-            Log.error(self, "Unable to restart Memcached", False)
-
-    @expose(hide=True)
     def clean_fastcgi(self):
-        if(os.path.isdir("/var/run/nginx-cache")):
+        if(os.path.isdir("/var/run/nginx-cache") and
+           os.path.exists('/usr/sbin/nginx')):
             Log.info(self, "Cleaning NGINX FastCGI cache")
             WOShellExec.cmd_exec(self, "rm -rf /var/run/nginx-cache/*")
+            WOService.restart_service(self, 'nginx')
         else:
             Log.error(self, "Unable to clean FastCGI cache", False)
 
     @expose(hide=True)
     def clean_opcache(self):
-        try:
-            Log.info(self, "Cleaning opcache")
-            wp = urllib.request.urlopen(" https://127.0.0.1:22222/cache"
-                                        "/opcache/opgui.php?reset=1").read()
-        except Exception as e:
-            Log.debug(self, "{0}".format(e))
-            Log.debug(self, "Unable hit url, "
-                      " https://127.0.0.1:22222/cache/opcache/"
-                      "opgui.php?reset=1,"
-                      " please check you have admin tools installed")
-            Log.debug(self, "please check you have admin tools installed,"
-                      " or install them with `wo stack install --admin`")
-            Log.error(self, "Unable to clean opcache", False)
+        opcache_dir = '/var/www/22222/htdocs/cache/opcache/'
+        if (os.path.exists('/usr/sbin/nginx') and
+                os.path.exists(
+                    '/var/www/22222/htdocs/cache/opcache')):
+            try:
+                Log.info(self, "Cleaning opcache")
+                if os.path.exists('{0}php72.php'.format(opcache_dir)):
+                    requests.get(
+                        "http://127.0.0.1/cache/opcache/php72.php")
+                if os.path.exists('{0}php73.php'.format(opcache_dir)):
+                    requests.get(
+                        "http://127.0.0.1/cache/opcache/php73.php")
+                if os.path.exists('{0}php74.php'.format(opcache_dir)):
+                    requests.get(
+                        "http://127.0.0.1/cache/opcache/php74.php")
+            except requests.HTTPError as e:
+                Log.debug(self, "{0}".format(e))
+                Log.debug(self, "Unable hit url, "
+                          " http://127.0.0.1/cache/opcache/"
+                          "php72.php,"
+                          " please check you have admin tools installed")
+                Log.debug(self, "please check you have admin tools installed,"
+                          " or install them with `wo stack install --admin`")
+                Log.error(self, "Unable to clean opcache", False)
 
 
 def load(app):
     # register the plugin class.. this only happens if the plugin is enabled
-    handler.register(WOCleanController)
+    app.handler.register(WOCleanController)
     # register a hook (function) to run after arguments are parsed.
-    hook.register('post_argument_parsing', wo_clean_hook)
+    app.hook.register('post_argument_parsing', wo_clean_hook)
